Fixed basic iOS build

+++ Source/Generic/ECTestPerformanceCounter.m
@@ -45,7 +45,7 @@ uint64_t dispatch_benchmark(size_t count, void (^block)(void)); // private API -
 	NSLog(@"%@ performing %ld iterations", label, (long)iterations);
 	uint64_t nanos = dispatch_benchmark(iterations, block);
 	NSTimeInterval average = (NSTimeInterval)nanos / 1000000000.0;
-	NSLog(@"%@ average over #%ld iterations was %lfs (%lldns)", label, iterations, average, nanos);
+	NSLog(@"%@ average over #%ld iterations was %lfs (%lldns)", label, (unsigned long)iterations, average, nanos);
 
 	return average;
 }
